database and entities redesigned

+++ e-commerce-backend/src/main/java/com/example/model/CartItem.java
@@ -6,13 +6,12 @@ import lombok.Setter;
 
 @Entity
 @Table(name = "cart_items")
-@Getter
-@Setter
+@Getter @Setter
 public class CartItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "cart_item_id")
-    private int cartItemId;
+    private Long cartItemId;
 
     @ManyToOne
     @JoinColumn(name = "user_id", nullable = false)
@@ -23,5 +22,6 @@ public class CartItem {
     private Product product;
 
     @Column(name = "quantity", nullable = false)
-    private int quantity;
+    private Integer quantity;
 }
+
